Rename modules and tests
