+++ backend/core/__init__.py
@@ -1,3 +1,3 @@
-from .settings import settings
+from .kodekloud_settings import settings
 
 __all__ = ["settings"]
