+++ backend/main.py
@@ -1,9 +1,9 @@
 from fastapi import FastAPI
-from app.routers import kodekloudreport
+from app.routers import kodekloud_report
 
 app = FastAPI(title="KodeKloud License API")
 
-app.include_router(kodekloudreport.router)
+app.include_router(kodekloud_report.router)
 
 
 @app.get("/")
