+++ backend/pytest.ini
@@ -1,0 +1,2 @@
+[pytest]
+python_files = kk_test_*.py
